fix Gamespy3 protocol
See #197. If the server (for instance Crysis Wars) has no connection to the Gamespy master server, it will not respond with a challenge in the expected format. However, if you leave out the challenge in the subsequent request, you can still query the server information.

This change checks if a challenge was successfully extracted from the challenge buffer and only then processes the challenge result. Otherwise the challenge result will be empty by default and thus simply omitted in the subsequent request.

Without this change, no Crysis Wars server will be able to queried otherwise, since the Gamespy Network is not existant anymore and Crysis Wars for example can be patched to use Qtracker instead.